undo files with pyllint info

+++ onnx/reference/ops/op_non_zero.py
@@ -1,6 +1,5 @@
-# Copyright (c) ONNX Project Contributors
-#
 # SPDX-License-Identifier: Apache-2.0
+# pylint: disable=W0221
 
 import numpy as np
 
@@ -9,6 +8,5 @@ from onnx.reference.op_run import OpRun
 
 class NonZero(OpRun):
     def _run(self, x):  # type: ignore
-        # Specify np.int64 for Windows x86 machines
-        res = np.vstack(np.nonzero(x)).astype(np.int64)
+        res = np.vstack(np.nonzero(x))
         return (res,)
